deploy_clean/requirements.txt
# Flask and core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Database drivers for PostgreSQL/RDS - AWS compatible
psycopg2-binary==2.9.9

# SQLAlchemy - compatible version
SQLAlchemy==2.0.23

# Security and environment
python-dotenv==1.0.0
Werkzeug==2.3.7

# Timezone handling
pytz==2023.3.post1

# Scheduler for background tasks
APScheduler==3.10.4

# Production server
gunicorn==21.2.0

# Additional utilities
requests==2.31.0 